Isolate Login login into the SophysLogin widget
